factor normalizePath out into separate module
Prepatory refactoring for the rest of this branch.
